Add initial migrations for Dragonarium application.
This commit includes the initial migrations for creating the database of the Dragonarium application. It includes the creation of several tables like 'Currencies', 'Dragons', 'Elements', 'Habitats', 'Items', 'DragonEggs', 'DragonElements', 'HabitatElements', 'ItemCurrencies'. It also contains their respective constraints and relations. This is necessary for storing and managing data in the application. The setting for ignoring certain namespace folders is included as well.