Remove migration for database initialization
The migration for initializing the database was removed. The "Initialize" migration sets up a number of tables related to the Dragonarium application. In terms of why it was removed, it is possible that this is being redesigned or that there are other ways of implementing the same functionality that have been decided to be better. Therefore the migration is no longer needed. It is also possible that the code has been moved to some other location or has been redundant and hence it has been removed.