"Add Config and NewGame scenes, update Main scene
Added two new scenes Config and NewGame with their setup. Config is a Node and NewGame is a control User interface (UI) for creating a new game, thus enhancing the users' experience. The Main scene is also updated by reducing the load steps from 8 to 7, and the HUDCntr instance under the Canvas is removed for better performance and efficiency."